first rough pass at client/server split
