+++ src/client/common.ts
@@ -1,0 +1,1 @@
+../common/index.ts
