+++ src/client/index.ts
@@ -17,4 +17,5 @@
 
 /** Client code. */
 
+export * from './common';
 export * from './api';
