+++ src/server/package.json
@@ -1,18 +1,20 @@
 {
-  "name": "federated-learning",
+  "name": "federated-learning-server",
   "version": "0.0.1",
   "description": "Hardware-accelerated JavaScript library for machine intelligence",
   "private": true,
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "module": "dist/federated-learning-client.esm.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/PAIR-code/federated-learning.git"
   },
   "license": "Apache-2.0",
   "scripts": {
-    "start": "node dist/src/server/server.js",
-    "dev": "ts-node src/server/audio_demo_server.ts",
-    "build": "tsc && cp demo/index.html dist/demo/",
-    "test": "ts-node node_modules/jasmine/bin/jasmine --config=jasmine.json",
+    "publish-local": "rm -rf dist/ && yarn && yarn build && yalc publish",
+    "dev": "ts-node audio_demo_server.ts",
+    "build": "tsc",
     "lint": "tslint -p . -t verbose"
   },
   "dependencies": {
@@ -23,7 +25,6 @@
     "@types/express-fileupload": "^0.1.1",
     "@types/node-fetch": "^2.1.1",
     "@types/socket.io": "^1.4.34",
-    "@types/socket.io-client": "^1.4.32",
     "@types/uuid": "^3.4.3",
     "bswap": "^1.2.0",
     "encoding-down": "^5.0.3",
@@ -42,12 +43,11 @@
     "@types/rimraf": "^2.0.2",
     "@types/seedrandom": "~2.4.27",
     "clang-format": "~1.2.2",
-    "jasmine": "^3.1.0",
-    "jasmine-core": "~3.1.0",
     "rimraf": "^2.6.2",
     "ts-node": "^6.1.0",
     "ts-node-dev": "^1.0.0-pre.26",
     "tslint": "~5.8.0",
-    "typescript": "2.7.2"
+    "typescript": "2.9.2",
+    "yalc": "^1.0.0-pre.22"
   }
 }
