+++ test/jasmine.json
@@ -1,5 +1,5 @@
 {
-  "spec_dir": "src",
+  "spec_dir": ".",
   "spec_files": [
     "**/*_test.ts"
   ]
