+++ test/tsconfig.json
@@ -9,7 +9,7 @@
     "declaration": true,
     "target": "es5",
     "lib": ["es2015", "dom"],
-    "outDir": "./dist",
+    "outDir": "./dist/test",
     "noUnusedLocals": true,
     "noImplicitReturns": true,
     "noImplicitThis": true,
@@ -21,7 +21,6 @@
     "experimentalDecorators": true
   },
   "include": [
-    "src/",
-    "demo/"
+    "./"
   ]
 }
